requirements.txt
llama_index==0.10.52
python-dotenv==1.0.1
